make V2 labels be just labels
