+++ .secrets.baseline
@@ -352,6 +352,32 @@
         "is_secret": false
       }
     ],
+    "osidb/migrations/0259_remove_flawlabel_insert_insert_and_more.py": [
+      {
+        "type": "Hex High Entropy String",
+        "filename": "osidb/migrations/0259_remove_flawlabel_insert_insert_and_more.py",
+        "hashed_secret": "a41e4af92bda7278b293b69f1a83b24ca051d2d3",
+        "is_verified": false,
+        "line_number": 58,
+        "is_secret": false
+      },
+      {
+        "type": "Hex High Entropy String",
+        "filename": "osidb/migrations/0259_remove_flawlabel_insert_insert_and_more.py",
+        "hashed_secret": "aeed0a320fb304b3835ad52cd53186c12a094e4a",
+        "is_verified": false,
+        "line_number": 73,
+        "is_secret": false
+      },
+      {
+        "type": "Hex High Entropy String",
+        "filename": "osidb/migrations/0259_remove_flawlabel_insert_insert_and_more.py",
+        "hashed_secret": "e508671d5b8996e548e870c414b50d230453988b",
+        "is_verified": false,
+        "line_number": 87,
+        "is_secret": false
+      }
+    ],
     "osidb/tests/endpoints/flaws/test_acknowledgments.py": [
       {
         "type": "Secret Keyword",
@@ -481,5 +507,5 @@
       }
     ]
   },
-  "generated_at": "2026-08-05T12:10:03Z"
+  "generated_at": "2026-08-06T14:52:39Z"
 }
