+++ osidb/models/tests/test_label.py
@@ -1,5 +1,5 @@
 """
-Tests for the new polymorphic label models (FlawLabelV2 hierarchy).
+Tests for the new polymorphic label models (FlawLabel hierarchy).
 
 These tests verify the polymorphic model behavior, type routing,
 and subclass-specific functionality.
@@ -12,13 +12,13 @@ from django.core.exceptions import ValidationError
 from django.db import IntegrityError
 from freezegun import freeze_time
 
-from osidb.models.flaw.label_v2 import (
+from osidb.models.flaw.label import (
     AliasLabel,
     BULabel,
     BULabelDefinition,
     CollaboratorLabel,
     CollaboratorLabelDefinition,
-    FlawLabelV2,
+    FlawLabel,
     ProductFamilyLabel,
     ProductFamilyLabelDefinition,
     WorkflowLabel,
@@ -98,12 +98,12 @@ class TestPolymorphicBasics:
         workflow = WorkflowLabel.objects.create(flaw=flaw, name="workflow-label")
         bu = BULabel.objects.create(flaw=flaw, name="bu-label", state=BULabel.State.NEW)
 
-        # Verify they're all FlawLabelV2 instances
-        assert isinstance(collab, FlawLabelV2)
-        assert isinstance(product, FlawLabelV2)
-        assert isinstance(alias, FlawLabelV2)
-        assert isinstance(workflow, FlawLabelV2)
-        assert isinstance(bu, FlawLabelV2)
+        # Verify they're all FlawLabel instances
+        assert isinstance(collab, FlawLabel)
+        assert isinstance(product, FlawLabel)
+        assert isinstance(alias, FlawLabel)
+        assert isinstance(workflow, FlawLabel)
+        assert isinstance(bu, FlawLabel)
 
         # Verify they're also their specific types
         assert isinstance(collab, CollaboratorLabel)
@@ -113,7 +113,7 @@ class TestPolymorphicBasics:
         assert isinstance(bu, BULabel)
 
     def test_polymorphic_query_returns_correct_subclass(self):
-        """Test that querying FlawLabelV2 returns the correct subclass instances"""
+        """Test that querying FlawLabel returns the correct subclass instances"""
         flaw = FlawFactory(embargoed=False)
         AffectFactory(flaw=flaw)
 
@@ -122,7 +122,7 @@ class TestPolymorphicBasics:
         WorkflowLabel.objects.create(flaw=flaw, name="workflow1")
 
         # Query the base class
-        labels = FlawLabelV2.objects.filter(flaw=flaw).order_by("name")
+        labels = FlawLabel.objects.filter(flaw=flaw).order_by("name")
 
         # Should get back the specific subclasses, not base class
         assert isinstance(labels[0], AliasLabel)
@@ -335,6 +335,49 @@ class TestProductFamilyLabel:
         # Verify the WorkflowLabel still exists
         assert WorkflowLabel.objects.filter(flaw=flaw, name="another-conflict").exists()
 
+    @pytest.mark.enable_signals
+    def test_update_relevance_exclusion_per_affect(self):
+        """
+        An exclusion on one affect must not suppress a definition
+        matched by a different affect.
+        """
+        ps_included = PsModuleFactory(name="included-module")
+        ps_excluded = PsModuleFactory(name="excluded-module")
+        PsUpdateStreamFactory(name="inc-stream", ps_module=ps_included)
+        PsUpdateStreamFactory(name="exc-stream", ps_module=ps_excluded)
+
+        ProductFamilyLabelDefinition.objects.create(
+            name="per-affect-test",
+            ps_modules=["included-module"],
+            ps_modules_exclude=["excluded-module"],
+        )
+
+        flaw = FlawFactory(embargoed=False)
+        AffectFactory(
+            flaw=flaw,
+            ps_module="included-module",
+            ps_component="comp-a",
+            ps_update_stream="inc-stream",
+        )
+        AffectFactory(
+            flaw=flaw,
+            ps_module="excluded-module",
+            ps_component="comp-b",
+            ps_update_stream="exc-stream",
+        )
+
+        # Force irrelevant so the assertion proves update_relevance
+        # restores it based on the included affect, not signal state.
+        ProductFamilyLabel.objects.filter(flaw=flaw, name="per-affect-test").update(
+            relevant=False
+        )
+
+        ProductFamilyLabel.update_relevance(flaw)
+
+        assert ProductFamilyLabel.objects.filter(
+            flaw=flaw, name="per-affect-test", relevant=True
+        ).exists()
+
 
 class TestAliasLabel:
     """Tests for AliasLabel model"""
@@ -528,7 +571,7 @@ class TestMixedLabelTypes:
         BULabel.objects.create(flaw=flaw, name="bu")
 
         # All should exist
-        labels = FlawLabelV2.objects.filter(flaw=flaw)
+        labels = FlawLabel.objects.filter(flaw=flaw)
         assert labels.count() == 5
 
         # Verify we get the right subclasses
